new Lucide Icon component maybe
