+++ app/components/Icon.tsx
@@ -1,16 +1,10 @@
 import * as React from "react"
 import { ComponentType } from "react"
-import {
-  Image,
-  ImageStyle,
-  StyleProp,
-  TouchableOpacity,
-  TouchableOpacityProps,
-  View,
-  ViewStyle,
-} from "react-native"
+import { StyleProp, TouchableOpacity, TouchableOpacityProps, View, ViewStyle } from "react-native"
+import * as LucideIcons from "lucide-react-native"
+import { LucideIcon } from "lucide-react-native"
 
-export type IconTypes = keyof typeof iconRegistry
+export type IconTypes = keyof typeof LucideIcons
 
 interface IconProps extends TouchableOpacityProps {
   /**
@@ -27,11 +21,6 @@ interface IconProps extends TouchableOpacityProps {
    * An optional size for the icon. If not provided, the icon will be sized to the icon's resolution.
    */
   size?: number
-
-  /**
-   * Style overrides for the icon image
-   */
-  style?: StyleProp<ImageStyle>
 
   /**
    * Style overrides for the icon container
@@ -51,14 +40,9 @@ interface IconProps extends TouchableOpacityProps {
  * - [Documentation and Examples](https://github.com/infinitered/ignite/blob/master/docs/Components-Icon.md)
  */
 export function Icon(props: IconProps) {
-  const {
-    icon,
-    color,
-    size,
-    style: $imageStyleOverride,
-    containerStyle: $containerStyleOverride,
-    ...WrapperProps
-  } = props
+  const { icon, color, size, containerStyle: $containerStyleOverride, ...WrapperProps } = props
+
+  const IconComponent: LucideIcon = LucideIcons[icon] as LucideIcon
 
   const isPressable = !!WrapperProps.onPress
   const Wrapper: ComponentType<TouchableOpacityProps> = WrapperProps?.onPress
@@ -71,44 +55,7 @@ export function Icon(props: IconProps) {
       {...WrapperProps}
       style={$containerStyleOverride}
     >
-      <Image
-        style={[
-          $imageStyle,
-          color && { tintColor: color },
-          size && { width: size, height: size },
-          $imageStyleOverride,
-        ]}
-        source={iconRegistry[icon]}
-      />
+      <IconComponent stroke={color || "#000"} width={size || 24} height={size || 24} />
     </Wrapper>
   )
-}
-
-export const iconRegistry = {
-  back: require("../../assets/icons/back.png"),
-  bell: require("../../assets/icons/bell.png"),
-  caretLeft: require("../../assets/icons/caretLeft.png"),
-  caretRight: require("../../assets/icons/caretRight.png"),
-  check: require("../../assets/icons/check.png"),
-  clap: require("../../assets/icons/clap.png"),
-  community: require("../../assets/icons/community.png"),
-  components: require("../../assets/icons/components.png"),
-  debug: require("../../assets/icons/debug.png"),
-  github: require("../../assets/icons/github.png"),
-  heart: require("../../assets/icons/heart.png"),
-  hidden: require("../../assets/icons/hidden.png"),
-  ladybug: require("../../assets/icons/ladybug.png"),
-  lock: require("../../assets/icons/lock.png"),
-  menu: require("../../assets/icons/menu.png"),
-  more: require("../../assets/icons/more.png"),
-  pin: require("../../assets/icons/pin.png"),
-  podcast: require("../../assets/icons/podcast.png"),
-  settings: require("../../assets/icons/settings.png"),
-  slack: require("../../assets/icons/slack.png"),
-  view: require("../../assets/icons/view.png"),
-  x: require("../../assets/icons/x.png"),
-}
-
-const $imageStyle: ImageStyle = {
-  resizeMode: "contain",
 }
