+++ package.json
@@ -54,6 +54,7 @@
     "expo-splash-screen": "~0.18.1",
     "expo-status-bar": "~1.4.4",
     "i18n-js": "3.9.2",
+    "lucide-react-native": "0.214.0",
     "mobx": "6.6.0",
     "mobx-react-lite": "3.4.0",
     "mobx-state-tree": "5.1.5",
